% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculator.R
\name{diversity_phylo}
\alias{diversity_phylo}
\title{Calculate phylo-diversity}
\usage{
diversity_phylo(PhyloTree, abundance_matrix, null = "taxa.labels", run = 999)
}
\arguments{
\item{PhyloTree}{A phylogenetic tree constructed by U.Phylomaker package.}

\item{abundance_matrix}{A species-plot abundance matrix.}
}
\value{
A tibble contains 3 columns: plot, NRI and NTI.
}
\description{
Calculate phylogenetic diversity using picante package, based on the construction of a phylogenetic tree and a species-plot abundance matrix.
}
