% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloCopilot.R
\name{get_phylo}
\alias{get_phylo}
\title{Build a APG IV-based phylotree}
\usage{
get_phylo(PlantList)
}
\arguments{
\item{PlantList}{The first column of the dataframe, or the entire vector, consists of standard Latin names recognized by the World Plants website.}
}
\value{
A phylogenetic tree constructed by U.PhyloMaker package.
}
\description{
Constructing a phylogenetic tree based on the APG IV classification system and the megatree backbone.
}
