% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhyloPloter.R
\name{plot_phylo}
\alias{plot_phylo}
\title{Plot a phylotree with node types}
\usage{
plot_phylo(PhyloTree, NoteColor = c("#f5c88f", "#b3c384", "#73a66b"))
}
\arguments{
\item{PhyloTree}{A phylogenetic tree constructed by U.Phylomaker package.}
}
\value{
A visualization of a phylogenetic tree with node types.
}
\description{
Visualizing a phylogenetic tree constructed by U.Phylomaker package, including resolution information for each node.
}
