% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Standardizer.R
\name{stand_taxon}
\alias{stand_taxon}
\title{WP-based species standardizer}
\usage{
stand_taxon(Plantlist, IsAuthor = F)
}
\arguments{
\item{Plantlist}{Inputs could be "vector" or "dataframe" as you like!}
}
\value{
A tibble contains 3 columns: AcceptedName, LoggedName and Equivalence.
}
\description{
Standardize the input list of plant species based on the World Plant Database with U.Taxonstand package.
}
\examples{
stand_taxon("Aa")
stand_taxon(c("Pinus massonlana","Cunninghamia lanceolata","Lindera glauca","Litsea cubeba","Sassafras tzumu"))
stand_taxon(data.frame("species"=c("Pinus massonlana","Cunninghamia lanceolata","Lindera glauca","Litsea cubeba","Sassafras tzumu")))
}
