% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APGClassifier.R
\name{get_taxon}
\alias{get_taxon}
\title{Get taxonomic info from APG IV}
\usage{
get_taxon(PlantList)
}
\arguments{
\item{PlantList}{dataframe or vector}
}
\value{
species, genus and family
}
\description{
This R package extracts genus and family information based on APG IV from space-corrected species names
}
