% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APGClassifier.R
\name{get_taxon}
\alias{get_taxon}
\title{Get taxonomic info from APG IV}
\usage{
get_taxon(PlantList)
}
\arguments{
\item{PlantList}{Inputs could be "vector" or "dataframe" as you like!}
}
\value{
A tibble contains 3 columns: species, genus and APG-IV family.
}
\description{
Extract genus and family information based on APG IV from space-corrected species names.
}
\examples{
get_taxon("Aa")
get_taxon(c("Pinus massonlana","Cunninghamia lanceolata","Lindera glauca","Litsea cubeba","Sassafras tzumu"))
get_taxon(data.frame("species"=c("Pinus massonlana","Cunninghamia lanceolata","Lindera glauca","Litsea cubeba","Sassafras tzumu")))
}
